+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Added privacy protection to option to send email without personal data by @hmccloy (#92)
 - Added list plugin option to set prefiltered locations (#68)
+- Added a new event to modify google for jobs data
 
 ### Fixed